moves all cf in a one script
