+++ components/cosmetic_filters/resources/data/cosmetic_filters_resources.grd
@@ -8,8 +8,7 @@
   </outputs>
   <release seq="1">
     <includes>
-      <include name="IDR_COSMETIC_FILTERS_SCRIPT" file="cosmetic_filters.js" type="BINDATA" />
-      <include name="IDR_COSMETIC_FILTERS_PUMP_SCRIPT" file="cosmetic_filters_pump.js" type="BINDATA" />
+      <include name="IDR_COSMETIC_FILTERS_SCRIPT" file="cosmetic_filters.js" type="BINDATA" compress="gzip"/>
     </includes>
   </release>
 </grit>
